rename component to InlineAlert
