--- packages/react-components/src/components/Alert/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./Alert";
-export type { AlertProps } from "./Alert";
