+++ packages/react-components/src/components/InlineAlert/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./InlineAlert";
+export type { InlineAlertProps } from "./InlineAlert";
